# optimize compat by hoisting the vnode handling func