This seems to significantly improve performance from a V8 Point of View for RTS

```
Text:
compat x 287.468 ops/sec ±2.71 (82 runs sampled)
current x 321.807 ops/sec ±2.38 (86 runs sampled)

SearchResults:    
compat x 735.515 ops/sec ±3.64 (83 runs sampled)
current x 867.643 ops/sec ±2.94 (86 runs sampled)

ColorPicker:      
compat x 3,190.427 ops/sec ±3.62 (89 runs sampled)
current x 4,169.366 ops/sec ±3.11 (89 runs sampled)
```

current is this, compat is the 10.13.1 release